+++ packages/nuxt/playground/app/app.vue
@@ -21,6 +21,15 @@ provide('navigation', navigation)
 
 @source "../content";
 
+@theme {
+  --font-geist: 'Geist', sans-serif;
+  --font-mono: 'Geist Mono', sans-serif;
+}
+
+:root {
+  @apply font-geist;
+}
+
 html[data-framework="nuxt"] .vue-only,
 html[data-framework="vue"] .nuxt-only {
   display: none;