+++ packages/nuxt/playground/nuxt.config.ts
@@ -1,3 +1,5 @@
+import pkg from '../package.json'
+
 export default defineNuxtConfig({
   future: {
     compatibilityVersion: 4
@@ -9,6 +11,21 @@ export default defineNuxtConfig({
   devtools: { enabled: true },
   compatibilityDate: '2024-07-09',
 
+  mdc: {
+    highlight: {
+      theme: {
+        dark: 'github-dark',
+        default: 'github-dark',
+        light: 'github-dark',
+      }
+    },
+  },
+
+  runtimeConfig: {
+    public: {
+      version: pkg.version
+    }
+  },
 
   routeRules: {
     '/getting-started/installation': { redirect: '/getting-started/installation/nuxt', prerender: false }