+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,26 @@
 # mockline
 
+## 0.9.1
+
+### Patch Changes
+
+- Fix themes build export
+- Updated dependencies
+  - @mockline/themes@0.9.1
+  - @mockline/types@0.9.1
+
+## 0.9.0
+
+### Minor Changes
+
+- Add prose components and complete docs refactor
+
+### Patch Changes
+
+- Updated dependencies
+  - @mockline/themes@0.9.0
+  - @mockline/types@0.9.0
+
 ## 0.8.0
 
 ### Minor Changes