+++ packages/themes/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @mockline/themes
 
+## 0.9.1
+
+### Patch Changes
+
+- Fix themes build export
+
+## 0.9.0
+
+### Minor Changes
+
+- Add prose components and complete docs refactor
+
 ## 0.8.0
 
 ### Minor Changes