+++ packages/types/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mockline/types
 
+## 0.9.1
+
+### Patch Changes
+
+- Fix themes build export
+
+## 0.9.0
+
 ## 0.8.0
 
 ### Minor Changes